requirements.txt
flask>=3.0
flask-limiter>=3.5
cryptography>=42.0
pyotp>=2.9
qrcode>=7.4
requests>=2.31
python-dotenv>=1.0
argon2-cffi>=23.1
pyjwt>=2.8
pillow>=10.0
gunicorn>=21.2
